% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{addNucleusPosition}
\alias{addNucleusPosition}
\title{addNucleusPosition}
\usage{
addNucleusPosition(object, segment.df, scale_file, tissue_positions.file)
}
\value{

}
\description{
addNucleusPosition
}
\examples{

}
\author{
Dieter Henrik Heiland
}
