% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfiler2Matlab}
\alias{CellProfiler2Matlab}
\title{CellProfiler2Matlab}
\usage{
CellProfiler2Matlab(mask_path, output_name)
}
\value{

}
\description{
CellProfiler2Matlab
}
\examples{

}
\author{
Dieter Henrik Heiland
}
