% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerCellDistance}
\alias{CellProfilerCellDistance}
\title{ImportCellProfiler data}
\usage{
CellProfilerCellDistance(data)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
