% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerDistancePlot}
\alias{CellProfilerDistancePlot}
\title{ImportCellProfiler data}
\usage{
CellProfilerDistancePlot(data, prefix, data.dist, exclude = NULL)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
