% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerImport}
\alias{CellProfilerImport}
\title{ImportCellProfiler data}
\usage{
CellProfilerImport(csv, type = NULL, var = "Mean", rename = T, rescale = T)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
