% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerImportObjectPositions}
\alias{CellProfilerImportObjectPositions}
\title{ImportCellProfiler data}
\usage{
CellProfilerImportObjectPositions(csv, type = NULL)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
