% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerPlotDistance}
\alias{CellProfilerPlotDistance}
\title{ImportCellProfiler data}
\usage{
CellProfilerPlotDistance(
  data.dist,
  image = 1,
  color = RColorBrewer::brewer.pal(8, "Dark2"),
  pt.size = 4,
  lt.size = 2,
  pt.alpha = 0.8,
  lt.alpha = 0.8,
  dist.threshold = 200
)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
