% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProfiler_family.R
\name{CellProfilerPlotDistanceAddLayer}
\alias{CellProfilerPlotDistanceAddLayer}
\title{ImportCellProfiler data}
\usage{
CellProfilerPlotDistanceAddLayer(
  p,
  image = 1,
  data,
  color.to,
  pt.size = 4,
  pt.alpha = 0.8,
  limit = c(0, 1)
)
}
\value{

}
\description{
ImportCellProfiler data
}
\examples{

}
\author{
Dieter Henrik Heiland
}
