% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{check_color_to}
\alias{check_color_to}
\title{check_color_to}
\usage{
check_color_to(object, color_to, color_by, max_length = NULL)
}
\value{

}
\description{
check_color_to
}
\examples{

}
\author{
Dieter Henrik Heiland
}
