% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{Create.ref.bins}
\alias{Create.ref.bins}
\title{Create.ref.bins}
\usage{
Create.ref.bins(object, bin.size)
}
\arguments{
\item{bin.size}{Size of Chromosomal bins}

\item{regions}{Input from the SPATA2 data.frame SPATA2::cnv_regions_df}
}
\value{

}
\description{
Create.ref.bins is a function to setup a reference data.frame for chromosomal bins
}
\examples{

}
\author{
Dieter Henrik Heiland
}
