% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMC_family.R
\name{createInternalReferenceDataset}
\alias{createInternalReferenceDataset}
\title{createInternalReferenceDataset}
\usage{
createInternalReferenceDataset(
  object,
  marker.list,
  q = 0.95,
  split_by = "ID.int"
)
}
\value{

}
\description{
createInternalReferenceDataset
}
\examples{

}
\author{
Dieter Henrik Heiland
}
