% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMC_family.R
\name{createReferenceDIM}
\alias{createReferenceDIM}
\title{createReferenceDIM}
\usage{
createReferenceDIM(object, marker.list, n = 500, booster = 0.5, batch = 50000)
}
\value{

}
\description{
createReferenceDIM
}
\examples{

}
\author{
Dieter Henrik Heiland
}
