% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_family.R
\name{createShell}
\alias{createShell}
\title{createShell}
\usage{
createShell(string, object.to.load, nameFromObject, wd = getwd())
}
\arguments{
\item{string}{The string contains a function that need to be run in a external shell}

\item{object.to.load}{The object that need to be processed (the function will save the object) ... name.obj ....}

\item{nameFromObject}{The name of the object in your environment as character ... "name.obj"...}

\item{wd}{The folder in which the object will be saved}
}
\value{

}
\description{
createShell
}
\examples{

}
\author{
Dieter Henrik Heiland
}
