% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_family.R
\name{DownScaleSeurat}
\alias{DownScaleSeurat}
\title{Run single-cell mapping}
\usage{
DownScaleSeurat(
  seurat,
  maintain_var = "cell_states",
  max = 10000,
  only_var = T,
  factor = 5,
  n_feature = 3000
)
}
\value{

}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
