% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getAdjacentSpots}
\alias{getAdjacentSpots}
\title{getAdjacentSpots}
\usage{
getAdjacentSpots(object, export.object = T, workers = 16, ram = 40)
}
\value{

}
\description{
getAdjacentSpots
}
\examples{

}
\author{
Dieter Henrik Heiland
}
