% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{getCNVStatistics}
\alias{getCNVStatistics}
\title{getCNVStatistics}
\usage{
getCNVStatistics(object, groups, across = NULL, sub.across = NULL)
}
\arguments{
\item{object}{SPATA2 object}
}
\value{

}
\description{
getCNVStatistics
}
\examples{

}
\author{
Dieter Henrik Heiland
}
