% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{getNucleusPosition}
\alias{getNucleusPosition}
\title{getNucleusPosition}
\usage{
getNucleusPosition(object)
}
\value{

}
\description{
getNucleusPosition
}
\examples{

}
\author{
Dieter Henrik Heiland
}
