% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_family.R
\name{getSingleCellDeconv}
\alias{getSingleCellDeconv}
\title{runRCTD cell-type deconvolution}
\usage{
getSingleCellDeconv(
  object,
  deconv_cell_types,
  spot_extension = NULL,
  multicore = T,
  flip.y = T,
  workers = 16,
  scale_factor = NULL
)
}
\value{

}
\description{
Run the a cell type deconvolution using the SpaceXR/RCTD algorithm from SPATA objects
}
\author{
Dieter Henrik Heiland
}
