% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getSpatialAdjacentRegression}
\alias{getSpatialAdjacentRegression}
\title{Spatial Adjacent Regression}
\usage{
getSpatialAdjacentRegression(
  object,
  feature1,
  feature2,
  smooth = F,
  normalize = F,
  workers = 16,
  ram = 40
)
}
\value{

}
\description{
Spatial Adjacent Regression
}
\examples{

}
\author{
Dieter Henrik Heiland
}
