% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_family.R
\name{getSubColors}
\alias{getSubColors}
\title{Run single-cell mapping}
\usage{
getSubColors(
  tab,
  class = "annotation_level_2",
  subclass = "annotation_level_4",
  pal = "Set3",
  max_pal = 12,
  random = T,
  seed = 200,
  into = "#EDEDED",
  add_n = 1
)
}
\value{

}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
