% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getSurroundedSpots}
\alias{getSurroundedSpots}
\title{getSurroundedSpots}
\usage{
getSurroundedSpots(object)
}
\value{

}
\description{
getSurroundedSpots
}
\examples{

}
\author{
Dieter Henrik Heiland
}
