% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{ggLayerGenSet}
\alias{ggLayerGenSet}
\title{plotCNV.RefMode}
\usage{
ggLayerGenSet(
  object,
  gene_set,
  y.pram = 1,
  pt.size = 1,
  pt.shape = 15,
  color = "black"
)
}
\arguments{
\item{object}{SPATA2 object}
}
\value{

}
\description{
plotCNV.RefMode
}
\examples{

}
\author{
Dieter Henrik Heiland
}
