% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{hlpr_bins}
\alias{hlpr_bins}
\title{hlpr_bins}
\usage{
hlpr_bins(Chr, start, end, bin.size)
}
\arguments{
\item{start}{Start position Chromosom}

\item{end}{End position Chromosom}

\item{bin.size}{Size of Chromosomal bins}

\item{chr}{Chromosome numeric}
}
\value{

}
\description{
hlpr_bins
}
\examples{

}
\author{
Dieter Henrik Heiland
}
