% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{hlpr_merge.pos}
\alias{hlpr_merge.pos}
\title{hlpr_merge.pos}
\usage{
hlpr_merge.pos(chr, start.pos, end.pos, ref.bin)
}
\arguments{
\item{start.pos}{Start position Chromosom}

\item{end.pos}{End position Chromosom}

\item{ref.bin}{Output of the Create.ref.bins() functions}

\item{Chr}{Chromosome numeric}
}
\value{

}
\description{
hlpr_merge.pos
}
\examples{

}
\author{
Dieter Henrik Heiland
}
