% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_family.R
\name{hlpr_smooth_shiny}
\alias{hlpr_smooth_shiny}
\title{Smooth variable spatially in mini-shiny-apps}
\usage{
hlpr_smooth_shiny(variable, coords_df, smooth_span)
}
\value{
Data.frame with the smoothed variable specified in \code{variable}.
}
\description{
Helper function to use independently (or in a pipe)
}
