% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNV_family.R
\name{importSNVmatrix}
\alias{importSNVmatrix}
\title{importSNVmatrix}
\usage{
importSNVmatrix(spata.obj, path.vcf, path.snv, count.filter = 1)
}
\value{

}
\description{
importSNVmatrix
}
\examples{

}
\author{
Dieter Henrik Heiland
}
