% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_family.R
\name{inferEnhancedPCA}
\alias{inferEnhancedPCA}
\title{inferEnhancedPCA}
\usage{
inferEnhancedPCA(object, cor = 8, ram = 10000)
}
\value{

}
\description{
inferEnhancedPCA
}
\examples{

}
\author{
Dieter Henrik Heiland
}
