% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferJuxtaposition}
\alias{inferJuxtaposition}
\title{Juxtaposition in stRNA-seq}
\usage{
inferJuxtaposition(
  object,
  source,
  source.type = "feature",
  target,
  target.type = "feature"
)
}
\arguments{
\item{source}{Variable define basline of state from which the distance is computet}

\item{source.type}{and target.type can be "feature", "gene" or "geneset"}

\item{target}{Variable "to" for distance computation}
}
\value{

}
\description{
Juxtaposition in stRNA-seq
}
\examples{

}
\author{
Dieter Henrik Heiland
}
