% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferSpatial.ac}
\alias{inferSpatial.ac}
\title{inferSpatial.mc}
\usage{
inferSpatial.ac(object, feature, radius = 5)
}
\value{

}
\description{
inferSpatial.mc
}
\examples{

}
\author{
Dieter Henrik Heiland
}
