% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferSpatial.mc}
\alias{inferSpatial.mc}
\title{inferSpatial.mc}
\usage{
inferSpatial.mc(object, P1, P2, n = 599)
}
\value{

}
\description{
inferSpatial.mc
}
\examples{

}
\author{
Dieter Henrik Heiland
}
