% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferSpotlight}
\alias{inferSpotlight}
\title{inferSpotlight}
\usage{
inferSpotlight(
  spata.obj,
  seurat.obj,
  feature,
  marker.genes = NULL,
  ntop = NULL,
  transf = "uv",
  min_cont = 0.09
)
}
\value{

}
\description{
inferSpotlight
}
\author{
Dieter Henrik Heiland
}
