% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferVectorFields}
\alias{inferVectorFields}
\title{inferVectorFields}
\usage{
inferVectorFields(
  object,
  df,
  parameter,
  dist.spot = 10,
  run.mcor = T,
  cor = 8,
  ram = 50000
)
}
\value{

}
\description{
inferVectorFields
}
\author{
Dieter Henrik Heiland
}
