% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_family.R
\name{MergeInferCNVSeurat}
\alias{MergeInferCNVSeurat}
\title{MergeInferCNVSeurat}
\usage{
MergeInferCNVSeurat(object, results = getwd())
}
\arguments{
\item{object}{Seurat Object}

\item{results}{The InferCNV results folder}
}
\value{

}
\description{
MergeInferCNVSeurat
}
\author{
Dieter Henrik Heiland
}
