% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plot2DInterpolation}
\alias{plot2DInterpolation}
\title{plot2DInterpolation}
\usage{
plot2DInterpolation(
  object,
  color_by,
  pt.size = 0.5,
  pt.alpha = 1,
  normalize = T,
  alpha2pred = T,
  smooth = F,
  smooth_span = NULL,
  addImage = F,
  Palette = NULL,
  pt_clrsp = "Reds",
  remove.pred = NULL,
  ...
)
}
\value{

}
\description{
plot2DInterpolation
}
\examples{

}
\author{
Dieter Henrik Heiland
}
