% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_family.R
\name{plotCNV}
\alias{plotCNV}
\title{plotCNV}
\usage{
plotCNV(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clr = "blue",
  ...,
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{object}{Seurat Object}

\item{results}{The InferCNV results folder}
}
\value{

}
\description{
plotCNV
}
\author{
Dieter Henrik Heiland
}
