% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{plotCNV.RefMode}
\alias{plotCNV.RefMode}
\title{plotCNV.RefMode}
\usage{
plotCNV.RefMode(
  object,
  across = NULL,
  sub.across = NULL,
  pt.size = 1.5,
  sample.order = T
)
}
\arguments{
\item{object}{SPATA2 object}
}
\value{

}
\description{
plotCNV.RefMode
}
\examples{

}
\author{
Dieter Henrik Heiland
}
