% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotCNVPoints}
\alias{plotCNVPoints}
\title{plotCNVPoints}
\usage{
plotCNVPoints(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clr = "blue",
  ...,
  of_sample = NA,
  verbose = NULL
)
}
\value{

}
\description{
plotCNVPoints
}
\examples{

}
\author{
Dieter Henrik Heiland
}
