% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotColorOverlap}
\alias{plotColorOverlap}
\title{plotColorOverlap}
\usage{
plotColorOverlap(
  object,
  feature1,
  feature2,
  pt_size = 3,
  get.map = F,
  two.colors = c("purple", "lightgreen"),
  negative.color = "darkgrey",
  col.threshold = 0.8,
  smooth = T,
  smooth_span = 0.2,
  normalize = T
)
}
\arguments{
\item{object}{SPATA2 object}

\item{feature1}{Character value. feature or gene to plot}

\item{feature2}{Character value. feature or gene to plot}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{get.map}{Logical.If TRUE plot only the color map}

\item{two.colors}{Character value, two colors}

\item{negative.color}{Character value, negative color}

\item{col.threshold}{Balance the colors}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().}

\item{normalize}{Normalize Values}
}
\value{

}
\description{
Plot overlap (Adopted from the Seurat Package)
}
\examples{

}
\author{
Dieter Henrik Heiland
}
