% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(object)
}
\arguments{
\item{object}{SPATA2 object}
}
\value{

}
\description{
plotCoverage
}
\examples{

}
\author{
Dieter Henrik Heiland
}
