% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_family.R
\name{plotEnhancedCluster}
\alias{plotEnhancedCluster}
\title{plotEnhancedCluster}
\usage{
plotEnhancedCluster(object, pt.size = 0.5, pt.alpha = 1, addImage = F)
}
\value{

}
\description{
plotEnhancedCluster
}
\examples{

}
\author{
Dieter Henrik Heiland
}
