% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_family.R
\name{plotEnhancedNucleusFeature}
\alias{plotEnhancedNucleusFeature}
\title{Enhanced Nucleus Feature}
\usage{
plotEnhancedNucleusFeature(
  object,
  genes,
  pt.size = 0.5,
  pt.alpha = 1,
  alpha2pred = F,
  addImage = F,
  Palette = NULL,
  pt_clrsp = "Reds"
)
}
\value{

}
\description{
Enhanced Nucleus Feature
}
\examples{

}
\author{
Dieter Henrik Heiland
}
