% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  object,
  across = NULL,
  gs,
  thr = c(0.8, 0.7, 0.7, 0.6, 0.5),
  nr.sample = NULL,
  plot.type = c("heatmap", "bar.comp", "point", "line", "stat"),
  size = 2,
  sample.type = "sample.type",
  color = "black",
  alpha = 1,
  feature = NULL
)
}
\value{

}
\description{
plotHeatmap
}
\examples{

}
\author{
Dieter Henrik Heiland
}
