% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotJuxtapositionSPATA}
\alias{plotJuxtapositionSPATA}
\title{plotJuxtapositionSPATA}
\usage{
plotJuxtapositionSPATA(
  object,
  feature.source = "NPC",
  feature.target = "OPC",
  color_by = "seurat_clusters",
  pt_alpha = 0.3,
  lt.alpha = 0.1,
  pt_size = 6,
  lt.size = 0.2,
  thr = 0.8,
  dist = 200
)
}
\value{

}
\description{
plotJuxtapositionSPATA
}
\examples{

}
\author{
Dieter Henrik Heiland
}
