% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNV_family.R
\name{plotSNVtype}
\alias{plotSNVtype}
\title{plotSNVtype}
\usage{
plotSNVtype(spata.obj, feature, type = "stack", order = NULL)
}
\arguments{
\item{type}{Can be "fill" or "stack}
}
\value{

}
\description{
plotSNVtype
}
\examples{

}
\author{
Dieter Henrik Heiland
}
