% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_family.R
\name{plotSPATAEnhancer}
\alias{plotSPATAEnhancer}
\title{plotSPATAEnhancer}
\usage{
plotSPATAEnhancer(
  object,
  color_by,
  normalize = T,
  smooth = F,
  smooth_span = NULL,
  pt.size = 0.5,
  pt.alpha = 1,
  alpha2pred = F,
  addImage = F,
  Palette = NULL,
  pt_clrsp = "Reds",
  ...
)
}
\value{

}
\description{
plotSPATAEnhancer
}
\examples{

}
\author{
Dieter Henrik Heiland
}
