% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_family.R
\name{plotSurfaceInteractiveSingleCell}
\alias{plotSurfaceInteractiveSingleCell}
\title{plotSurfaceInteractiveSingleCell}
\usage{
plotSurfaceInteractiveSingleCell(object)
}
\value{

}
\description{
plotSurfaceInteractiveSingleCell
}
\examples{

}
\author{
Dieter Henrik Heiland
}
