% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotSurfaceMixed}
\alias{plotSurfaceMixed}
\title{plotSurfaceMixed}
\usage{
plotSurfaceMixed(
  object,
  Mixed,
  mixed.colors,
  smooth = T,
  smooth_span = NULL,
  pt_size = 2.5,
  pt_alpha = T,
  display_image = F
)
}
\arguments{
\item{object}{SPATA2 object}

\item{Mixed}{Character value. Specifies the order and features that should be vizualized}

\item{mixed.colors}{colors from the palettes: colourvalues::color_palettes()}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_alpha}{Logical. If TRUE, the feature will be used as alpha}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample is displayed underneath the plot.}
}
\value{

}
\description{
Plot multiple parameter from the feature data.frame
}
\examples{

}
\author{
Dieter Henrik Heiland
}
