% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotVectorFields}
\alias{plotVectorFields}
\title{plotVectorFields}
\usage{
plotVectorFields(
  VF,
  parameter,
  pt.size = 6,
  pt.alpha = 0.8,
  color.extern = NULL,
  skip = 1
)
}
\value{

}
\description{
plotVectorFields
}
\examples{

}
\author{
Dieter Henrik Heiland
}
