% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimred_family.R
\name{rnmf}
\alias{rnmf}
\title{rnmf}
\usage{
rnmf(
  A,
  k,
  alpha = 0,
  init = "ica",
  n.cores = 1,
  loss = "mse",
  max.iter = 500,
  ica.fast = F
)
}
\value{

}
\description{
rnmf
}
\examples{

}
\author{
Dieter Henrik Heiland
}
