% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_family.R
\name{run.SNN.stability}
\alias{run.SNN.stability}
\title{Juxtaposition in stRNA-seq}
\usage{
run.SNN.stability(
  object,
  assay = "RNA",
  reduction = "pca",
  dims = 1:30,
  resolution = seq(from = 0.1, to = 1.5, by = 0.025),
  cluster_id = "new",
  algorithm = 1,
  plot = T
)
}
\value{

}
\description{
Juxtaposition in stRNA-seq
}
\examples{

}
\author{
Dieter Henrik Heiland
}
