% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{runCNV.Coverage}
\alias{runCNV.Coverage}
\title{Create.ref.bins}
\usage{
runCNV.Coverage(object)
}
\arguments{
\item{object}{SPATA2 object}

\item{regions}{Input from the SPATA2 data.frame SPATA2::cnv_regions_df}

\item{bin.size}{Size of Chromosomal bins}
}
\value{



}
\description{
Create.ref.bins is a function to setup a reference data.frame for chromosomal bins

runCNV.Coverage
}
\examples{


}
\author{
Dieter Henrik Heiland

Dieter Henrik Heiland
}
