% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_family.R
\name{runCNV.Normalization}
\alias{runCNV.Normalization}
\title{runCNVNormalization}
\usage{
runCNV.Normalization(
  object,
  window.k = 10,
  noise = 0.035,
  coverage.model = NULL
)
}
\arguments{
\item{object}{SPATA2 object}
}
\value{

}
\description{
runCNVNormalization
}
\examples{

}
\author{
Dieter Henrik Heiland
}
