% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_family.R
\name{runFullclustervalidation}
\alias{runFullclustervalidation}
\title{run.full.cluster.validation (for Spatial Paper Beta Version!!!!!)}
\usage{
runFullclustervalidation(seuratObj, plot.folder)
}
\value{

}
\description{
run.full.cluster.validation
}
\examples{

}
\author{
Dieter Henrik Heiland
}
