% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_family.R
\name{runMapping}
\alias{runMapping}
\title{Run single-cell mapping}
\usage{
runMapping(
  object,
  scDF,
  ref.new,
  cell_type_var = "annotation_level_4",
  model = c("oGA", "BF"),
  subset_ref = T,
  only_var = T,
  n_feature = 3000,
  max_cells = 10000,
  AE_norm = F,
  dropout = 0,
  bottleneck = 16,
  activation = "relu",
  layers = c(128, 64, 32),
  epochs = 20,
  multicore = T,
  workers = 16,
  iter = 200,
  iter_GA = 20,
  nr_mut = 2,
  nr_offsprings = 7,
  cross_over_point = 0.5,
  ram = 60
)
}
\arguments{
\item{object}{SPATA2 object}

\item{scDF}{Data.frame; Output of the getSingleCellDeconv()}

\item{ref.new}{Seurat Object; Reference dataset used for runRCTD()}

\item{cell_type_var}{Character value; The col of the Seurat meta data indicating the cell type annotations}

\item{model}{Character value; Model "BF" randomly select cell compositions and select the best match. Model "oGA" will perform a conditional genetic algorithm to find the best match.}

\item{subset_ref}{Logical, if TRUE, Seurat object will we downsized to increase speed.}

\item{only_var}{Logical, if TRUE only consider variable genes for mapping}

\item{n_feature}{Integer value: Number of variable genes}

\item{max_cells}{Integer value: Number of cells for downsampling}

\item{AE_norm}{Logical, if TRUE use an autoencoder for data integration and normalization (recommended)}

\item{multicore}{Logical, if TRUE use multicore}

\item{workers}{Integer, Number of cors}

\item{iter}{Integer: For model BF: Number of random spot compositions; For model oGA size of initial population}

\item{iter_GA}{Integer: Number of iterations of the oGA model}

\item{nr_mut}{Integer: Number of mutations}

\item{nr_offsprings}{Integer: Number of offsprings}

\item{cross_over_point}{Numeric value: Percentage of cross-over cutt-off}

\item{ram}{Integer: GB of ram can be used for multicore session}
}
\value{

}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
