% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimred_family.R
\name{runNMF}
\alias{runNMF}
\title{runNMF}
\usage{
runNMF(
  object,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  nfactors = 20,
  rescale = TRUE,
  reduction.name = "NMF",
  reduction.key = "factor_",
  n.cores = NULL,
  order.by.spcor = FALSE,
  sort.spcor.by.var = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}
}
\value{

}
\description{
runNMF adopted from STutility
}
\examples{

}
\author{
Dieter Henrik Heiland adopted from STutility
}
