% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_family.R
\name{runPCAprojection}
\alias{runPCAprojection}
\title{run.full.cluster.validation (for Spatial Paper Beta Version!!!!!)}
\usage{
runPCAprojection(
  PCA,
  matrix.input,
  matrix.transfer,
  Dim = 2,
  var = 400,
  epochs = 50
)
}
\value{

}
\description{
run.full.cluster.validation
}
\examples{

}
\author{
Dieter Henrik Heiland
}
