% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_family.R
\name{runSpatialRegression}
\alias{runSpatialRegression}
\title{runSpatialRegression}
\usage{
runSpatialRegression(object, features, model, smooth = F, normalize = F)
}
\value{

}
\description{
runSpatialRegression
}
\examples{

}
\author{
Dieter Henrik Heiland
}
