% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_family.R
\name{runVectorFields}
\alias{runVectorFields}
\title{runVectorFields}
\usage{
runVectorFields(
  object,
  features,
  cut_off = NULL,
  normalize = T,
  smooth = T,
  smooth_span = NULL,
  dist.spot = 10,
  run.mcor = T,
  workers = 8,
  ram = 50
)
}
\arguments{
\item{object}{SPATA2 object}

\item{features}{Character value. Can be a numeric feature from the feature data, a gene or a pathway}

\item{cut_off}{Numeric, cut off low expression spots}

\item{normalize}{Normalize Values}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing. Given to argument span of stats::loess().}

\item{dist.spot}{The distance for compute the vector (min is the distance of spots)}

\item{run.mcor}{Logical, if TRUE use feature multicore}

\item{workers}{Cores to use for multicore}

\item{ram}{Numeric, GB of ram storage used for multicore}
}
\value{

}
\description{
The function computed teh vector direction of gene expression gradients along a given feature
}
\author{
Dieter Henrik Heiland
}
