% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_family.R
\name{getSingleCellDeconv}
\alias{getSingleCellDeconv}
\title{runRCTD cell-type deconvolution}
\usage{
getSingleCellDeconv(
  object,
  deconv_cell_types,
  spot_extension = NULL,
  multicore = T,
  flip.y = T,
  workers = 16,
  scale_factor = NULL
)
}
\arguments{
\item{object}{SPATA2 object}

\item{deconv_cell_types}{Character value of all celltypes (from the feature data.frame) should be used}

\item{spot_extension}{Numeric value between 0 and 0.7. Specifies extension of the spot radius to include more cells. For example, a spot extension of 0.2 will increase the spot radius of 20\%)}

\item{multicore}{Logical. If TRUE will be run on all cores (recommended)}

\item{flip.y}{Logical. If TRUE y axis will be flipped}

\item{scale_factor}{Numeric value, if adjustment of the SPATAwrappers::getNucleusPosition() is required}
}
\value{

}
\description{
Run the a cell type deconvolution using the SpaceXR/RCTD algorithm from SPATA objects
}
\author{
Dieter Henrik Heiland
}
