% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferJuxtaposition}
\alias{inferJuxtaposition}
\title{Juxtaposition in stRNA-seq}
\usage{
inferJuxtaposition(object, genset = NULL, feature = NULL)
}
\value{

}
\description{
Juxtaposition in stRNA-seq
}
\examples{

}
\author{
Dieter Henrik Heiland
}
