% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferSpatial.plot}
\alias{inferSpatial.plot}
\title{inferSpatial.mc}
\usage{
inferSpatial.plot(object, feature, plot = T, radius = 5)
}
\value{

}
\description{
inferSpatial.mc
}
\examples{

}
\author{
Dieter Henrik Heiland
}
