% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_family.R
\name{inferSpotlight}
\alias{inferSpotlight}
\title{inferSpotlight}
\usage{
inferSpotlight(spata.obj, seurat.obj, feature, marker.genes = NULL)
}
\value{

}
\description{
inferSpotlight
}
\author{
Dieter Henrik Heiland
}
