% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotJuxtaposition}
\alias{plotJuxtaposition}
\title{plotJuxtaposition}
\usage{
plotJuxtaposition(dist, points = T, span)
}
\value{

}
\description{
plotJuxtaposition
}
\examples{

}
\author{
Dieter Henrik Heiland
}
