% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotStreamlines}
\alias{plotStreamlines}
\title{plotStreamlines}
\usage{
plotStreamlines(
  VF,
  parameter,
  size.arrow = 1,
  surface = T,
  grid = c(50, 50),
  gamma.u = 0.2,
  gamma.v = 0.2,
  arrow.angle = 25,
  arrow.length = 0.6,
  L = 100,
  color.extern = NULL,
  res = 1,
  pt.size = 6,
  pt.alpha = 0.8,
  skip = 2.5,
  xwrap = NULL,
  ywrap = NULL
)
}
\value{

}
\description{
plotStreamlines
}
\examples{

}
\author{
Dieter Henrik Heiland
}
