% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_family.R
\name{runBayesSpace}
\alias{runBayesSpace}
\title{BaySpace Cluster}
\usage{
runBayesSpace(
  object,
  pathToOuts,
  Spatial.enhancer = T,
  max.cluster = 13,
  return.model = T,
  empty.remove = F
)
}
\value{

}
\description{
BaySpace Cluster
}
\author{
Dieter Henrik Heiland
}
