% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_family.R
\name{runSpatialRegression}
\alias{runSpatialRegression}
\title{runSpatialRegression}
\usage{
runSpatialRegression(object, features, model, smooth = F, normalize = F)
}
\arguments{
\item{object}{SPATA2 object}

\item{features}{Character of all factors to be correlated (features/genes/GeneSets)}

\item{model}{The model to be used: classical:Pearson Correlation; lmSLX, lagsarlm, errorsarlm, CCA or dist}

\item{smooth}{Logical, if TRUE spatial smooth values}

\item{normalize}{Logical, if TRUE normalize values}
}
\value{

}
\description{
runSpatialRegression
}
\author{
Dieter Henrik Heiland
}
